Modify shared preferences.
Combined preferences in GameActivity to preferences to
MainActivity. Removed the need for saving button size. Modify
strings.xml
